added test and fixed error message, code cleaning

+++ src/main/java/it/phibonachos/andromeda/types/PositiveNum.java
@@ -5,6 +5,6 @@ import java.util.Optional;
 public class PositiveNum<NT extends Number> extends NumericConstraint<NT> {
     @Override
     public Boolean validate(NT number) {
-        return super.validate(number) && Optional.ofNullable(number.intValue() > 0 ? true : null).orElseThrow(NullPointerException::new);
+        return super.validate(number) && Optional.ofNullable(number.doubleValue() > 0 ? true : null).orElseThrow(NullPointerException::new);
     }
 }
